Choose your own cascade
Node/Link cascade files have been designed for simple network
(cascade-simple.xlsx) and the default complex one (cascade.xlsx).
Run model.py after choosing which to load with settings.
Note that 2000 to 2030 in timesteps of 0.25 take about half a minute on
my machine to run for two populations in the full cascade.
That's not promising, given what is left to be implemented.