Reorganising linkpar name loading
There is plenty of reorganisation left to do.
Also note that some xlsxwriter formulae will not 'recalculate' until the
sheet is opened up and saved in Excel, meaning that equation values have
to be expressly written from the Python end if they are to be recognised
as anything other than zero.
This has only been an issue thus far for Excel formulae that reference
other sheets.