Time-dependent parameters working
Have created a modded data sheet to demonstrate reinfection rate
oscillating between 1 and 0 for population two in the model.
Run test.py as is to compare plots.
Note that any value under assumption in the spreadsheet trumps all
time-dependent data. This should probably be explained in an information
box at the top of the sheet, when somebody is able to develop those.
Also, parameter interpolation must be slowing down the model somewhat as
it is 4 times slower on my machine now that everything is PCHIP
interpolated.
Will consider including an if clause for constants to avoid
interpolation.