Spreadsheet code shifted to databook file
Project methods for making/loading databooks are now wrappers for
function stored in the databook.py file.
This allows spreadsheet specific constants to be shared across functions
and raises the debate of whether databook metadata is best stored in a
Settings object or the databook scope. To be determined later.