Start of aging transitions in databook
Foundations of spreadsheet constructed. Lots left to do.
Very easy to have mismatching metadata in terms of row and column
titles, as well as placement, so might need to start developing databook
metadata dictionary in settings.
Also, death compartments should be flagged as they will not be part of
aging.