Prevalence and infection number plotting
Custom characteristics are now calculated at the end of a model run and
subsequently plotted in the temporary plotting script within the Project
object.
Characteristics in the cascade workbook can be tagged to 'Plot
Percentage', so that at plotting time, a fraction may be multiplied by
100.
This thus allows for infections and prevalence to be calculated, as well
as metrics such as living people and so on.