Example for different prevalence and infections
Test script test.py loads the modded project data sheet and plots custom
characteristics for the four-pop example involving the activation of
deaths only for juvenile prisoners.
As the juvenile prisoners die off or age into the adult prisoner
category, the number of total infections decreases, but prevalence rises
as everybody that is left alive progresses to active infection.
Proof that calculation of custom characteristics appears to work.