Commenting lines for timestep investigation
To change transition values rigorously from one timestep to another, it
is necessary to produce an operator matrix A out of all transitions
values and then take a fractional power of it, i.e. A^dt.
Doing this rigorously will hit performance especially during
optimisations, so it will need to be a huge conceptual refactoring down
the road.
For now, this issue can mostly be avoided by working with a timestep of
one year, so the exploration is currently commented out in test.py.
Getting formats working is a higher priority.