Significant change to compartment indexing
A ModelCompartment index is now expected to be a tuple, with first
element as its corresponding ModelPopulation index.
Accordingly, links will contain tuples as their index_to and index_from
attributes, allowing for inter-population links to be stored in the same
list as intra-population links and still be distinguished.
All transfer (i.e. migration) links are thus now stored in their
'source' population. This will allow all transitions to be applied
simultaneously.
However, this raises a few potential issues in terms of the getComp()
function, so function printLinkVars() is currently commented out.
Function stepCascadeForward() also forces a crash if any
inter-population links are found in the relevant population.