Fraction and Probability reconciliation
First success in dealing with transition rates that are given either as
a yearly 'fraction' of compartment size or a yearly independent
'probability' of a compartment member moving.
Conversion from one size of timestep dt to another also appears to be
working.
So, as an example, two outflows with yearly movement probabilities of
0.2 and 0.1 will produce dt-based values commensurate with actual yearly
fractions of ~1.9 and ~0.09, whereas two outflows with yearly movement
fractions 0.2 and 0.1 will produce values commensurate with actual
yearly fractions 0.2 and 0.1.
It follows that a fraction of 0.7 and a probability of 0.4 will
correspond to values commensurate with fractions 0.7 and ~0.2. But if
both are entered as fractions, the numerical conversion procedure should
crash (i.e. negative population problem).
This is a complicated concept and extremely damaging to performance.
Requires more thought to maintain this level of rigour.