Cleaned up transition conversion process
Test script test.py shows how infection/vaccination yearly fractions can
be maintained if transition-test databook format is set as 'fraction'.
However, population 3 in the transition-test databook has a very simple
setup, with no population inflows.
For population 1 and 2, with aging set as a 0.1 yearly fraction, actual
yearly movement is at ~0.12 for a dt of 0.25. This needs some thought to
remedy.