Characteristics propagated to parset
ParameterSet.pars is now subdivided into 'cascade' parameters and
'characs' parameters, the latter of which contains pop sizes and
prevalences.
Model even currently interpolates pop size to the start year and uses
that value as the starting susceptibles value.
Note: Collection compartments are currently not being counted as part of
standard cascade characteristics, despite having people in them. This
will be dealt with soon when collection compartments propagate
immediately...