Upgraded recursive characteristics validation
Rather than relying on definition order to mean something for
characteristic entry-points (especially considering that this order is
already restricted, with earlier characteristics unable to include later
characteristics), an alternate form of validation has been developed.
If a characteristic 'includes' the entry-point of another characteristic
and that characteristic includes the entry-point of the original (or
some more complex cyclical reference), the program throws a
recursion-limit exception.
This enforces a valid ordering when calculating initial values.