Initialisation with specified characteristics complete
Normalised characteristics (e.g. prevalence) can now be used to seed a
model, provided that the denominator (e.g. number of living people) is
also used to seed the model.
Run test.py with this commit to see default values for living-people,
active infections, latent infections and smear negative prevalence
(clearly interdependent) converted into appropriate initial values for a
model run.
Muck around with the 'Cascade Characteristics' sheet of cascade.xlsx to
either propagate alternative combinations or trigger numerous
settings.py validation checks.
On that point, it is not obvious to a user what is allowed/disallowed in
defining custom characteristics, so documentation will be required
eventually for devs.