Enabled hiding characteristics in databook
The 'Databook Order' column for both 'Cascade Characteristics' and
'Transition Parameters' now treats negative values as a tag to not print
out a data-entry block for the corresponding characteristic/parameter.
This will crash for a characteristic with an entry point, given that the
model cannot be initialised without values inserted in a databook. But
normal characteristics are nothing more than calculated outputs, so this
is not a problem for those.
Negative order is less advised for parameters as they are all used to
drive the model, but parameters missing from the databook are still
fine. They will simply be filled with relevant defaults or set to zero.
Note: Changes to a wide variety of files in this commit are cosmetic and
to do with OptimaException printouts.