Parameters/Characteristics written to custom sheets
The order of the concatenated parameter/characteristic odicts is not
being preserved, so this will need to be fixed.
However, data entry blocks for parameters/characteristics are being
written to databook sheets of choice.
If all parameters/characteristics are customised (or not printed to the
databook at all), there is no need for the standard sheets and they are
excluded.
Next up, loading...