Updated cascade for infectious prevalences
Infection rates are dependent on six prevalences, corresponding to the
coupled undiagnosed and diagnosed compartments of the six smear-strain
combinations.
It is assumed people on treatment are not infectious, in some cases due
to inherent quarantine for virulent strains, but this can be amended
later.
Characteristics have thus been refined to prepare for these
dependencies.