Simplest dependency case seemingly working
Run test.py for this commit to see slow-latent diagnosis rate be equal
to another parameter, as defined in cascade.xlsx.
Note: Performance does slow down quite a bit even for this simple
parameter-as-a-function-of-parameters step, which indicates complete
array pre-calculation should be done if there is no recursive
dependencies on characteristics. However, this is finnicky and can be
delayed.