Functional parameters calculated in proper order
Handled a situation where a tagged parameter may depend on an untagged
parameter that then depends on another untagged parameter. More
practically; a transition rate that may be another transition rate times
a suppression factor.
All parameters that are functions of other parameters are now listed as
par_funcs in settings.py, which helps in proper-ordered evaluation after
characteristics in Model's updateDependencies().