Edits pre-merge
Plotting outputs are currently hardcoded in plotting.py. Thus need to
swap comments in order to run either diary files or test.py.
Other than that, these edits seem to maintain functionality on both
branches.
The cascade.xlsx in the Optima folder is more advanced than the one in
the data folder, but both are being kept separate for the moment, with
data/cascade.xlsx only given an 'alive' entry point to actually run the
model.