Rearranging data files
Although having a data folder in parallel to the the optima folder that
contains test.py (i.e. "../data/") may lead to odd behaviour for custom
back-end user scripts, files have nonetheless been reorganised to adhere
to this structure for the time being. It can be reverted or modified
later with ease.
The diary folder will presumably be deleted from any eventual deployment
branch, so its scripts and outdated data file references have been
contained/maintained within the folder.
The data folder currently stores only cascade.xlsx and
cascade-simple.xlsx, although the latter is there for debugging purposes
and will be moved when designing a tutorial for cascade functionality in
the distant future.