Isolated countrybook settings as custom method
While Settings.loadCascadeSettings() is not powerful enough to produce
the desired user-viewed Belarus spreadsheet, it is necessary to store
country-book framework settings.
However, given that this is substantial hard-coding and is intended to
be replaced by cascade.xlsx functionality in the long term, it is
isolated to its own method for the moment.