Added y_factor (and equivalent field 'scale_factor' in Compartment) in model.py, as well as in required databook, settings, and parameter objects. Also allowed for flag indicating that transition should not be rescaled (during manual or automated caliibration), and example spreadsheet in cascade-simple-calibration.xlsx.
Effects:
- cascade spreadsheet: Parameters sheet can now contain 'Calibrate?' column
which, if value == Settings.DO_NOT_SCALE [== -1] will not attempt to rescale
the variable during automated calibration.
- as a default, the scale_factor / y_factor is Settings.DEFAULT_YFACTOR == 1.
- this is propagated through to link objects
- during model, if the link for a cascade rate is not Settings.DO_NOT_SCALE,
then the effective value is calculated as rate*y_factor.

This was tested before committing by inserting into model.py (in _calculateDPops:
line 418-9)
if ti < 20:
    link.scale_factor = 0.5 # or another small amount
and using the code example in diary/161123_runManualCalibration.py for
generating the plots. A kink should appear in the traces for some but
not all of the properties, depending on the values set in
cascade-simple-calibration.xlsx