Iteration check for junction processing
Currently using recursion_limit in settings to mark when infinite loop
for junctions is suspected.
Technically the limit should be much lower in general, perhaps ~10, but
this is not actually a big performance deal as the limit is designed to
crash the process anyway.
Still need to evaluate tolerances and negative checks though.