Major fix to model initialisation
The algorithm for converting characteristic starting values to
compartment starting values is spelled out in model.py.
Basically, if compartment 'a' is not an entry-point; a = 0.
If compartment 'a' is the entry-point of characteristic 'A', including
'a', 'b', 'c', then; a = A - b - c.
If 'A' only includes 'a', then compartment 'a' initialises with the
interpolated value of characteristic 'A'.
If 'A' includes more than 'a', the values of those compartments must
first be 'calculated' before being used.
So, if calculations are ordered, a = A, b = B - a - c, and c = C - a; b
can subtract a but not c until the next loop after c is calculated.
Complicated, but see model.py for more detail.