Hotfix to prevalence initialisation
Forgot to include denominator multiplications when fixing
initialisations yesterday. This has been rectified.
Of course, predictable bad things will happen if using a prevalence as a
denominator, as intended.
Also, cascade loading has been updated so that the Plot Value column can
be shunted after the 'includes' section of the characteristics sheet.
My preference is to remove the plot value column altogether and default
to no characteristics being plotted.