Cascade for careful review with UCL
Tentative completion of cascade.xlsx. Compartments and transitions are
effectively set in stone for Belarus.
Characteristics and parameters need combined review and discussion.
Notifications are currently better treated as parameters than
characteristics and inform partitioning.
Because partitioning involves junctions and these outflows need
immediate calculation, an extra dependency-update is run before the
initial junction-processing, so as to update ratios with user-input
notification values.
As for characteristics sheet, all compartments (except death and birth)
are entry points for some characteristic. Whether number or prevalence
is more important will be discussed.
Have also demonstrated custom output characteristics such as the
efficiency of testing and treatment in country, perhaps to answer
90-90-90 questions.