Rescales initial compartment values during autofitting.
Note that the step sizes for altering compartments are applied as
factors rather than additive step sizes (this can be improved). 

As a necessity, compartment sizes are only altered for characteristics
that have entry_point values associated with them. To prevent people
disappearing (or being randomly created), characteristics now also has a
'Calibrate?' column (as for the link parameters cascade values) and
entry points associated with total population (e.g. alive = all people,
with susceptible compartment as an entry point) is recommended as a "do
not calibrate this compartment".

An example databook is supplied, with the target link parameters and
initial compartments is also supplied. 

Note that autofit is still fragile - it requires a decent starting value
and can get stuck on "bad" solutions, but it is now useable.