Divide migration 'number' by valid compartments
Settings now stores (the resettable) attribute num_transfer_nodes.
When cascade settings are loaded, compartments are summed up if they are
not birth/death/junction related.
This gives an appropriate normalisation factor when a number of people
migrate from one population to another, assuming an equal split between
compartments.
This is not completely satisfying as the number of sus may be much
larger than those with active TB. So, thus far, this is a stopgap
solution.