Marking links that are transfers
For convenience, links that are produced as part of transfers are now
marked with a True is_transfer attribute.
During model calculations, any transfer links that have a number format
are scaled according to the proportion of people currently in the
compartment versus 'alive'.
This should give more accurate transfer behaviour, but the hard-coding
needs to be handled.
As a side note, all code involving num_transfer_nodes has been commented
out.