Default 'alive' characteristic now always created
When loading a cascade, all compartments that are not tagged birth or
dead (with junctions being ok) become an inclusion list for a default
characteristic; i.e. think population count.
If the user defines a characteristic that identically contains this
inclusion list, e.g. 'alive', then this overwrites the default (possibly
used for calibration/initialisation).
If not, then 'std_norm' (currently) becomes the substitute 'alive'.
This is required for situations where implicit definitions (e.g. in
cascade.xlsx) are not enough. User-defined inter-population transfers
are currently done with respect to the entire population count, so
model.py now makes sure that normalisation can be explicitly done.
The code in this commit should hopefully now allow for compartment-size
proportional migrations.