Dependencies handled for default characteristic
Characteristics included by other characteristics are not marked as
dependencies if the superset is merely an outcome. After all,
calculateOutputs() in model.py handles everything internally and
efficiently at the very end of a model run.
Only if the superset is identified as a parameter dependency (i.e. in a
parameter's custom function), then do all the subset characteristics
also get marked as dependencies (i.e. to be calculated per timestep).
This was a problem for the default pop-count characteristic which isn't
technically a parameter dependency, but is still required for explicit
calculations. Marking it as a dependency but ignoring the subsets leads
to testing issues.
This commit should thus treat the issue, marking the pop-count
characteristic and all subsets as dependencies, if not done so by the
end of parameter-sheet parsing.