Attempts for Demographic Fit
Calculations mentioned in the e-mail are in the databook-belarus sheet,
with the perfect scenario of 0 difference left to see and compare
against the original values below.

Updated the SP Untreated Death rates in the cascade sheet according to
Lara’s e-mail.

Line added to project/run_belarus.py by David to change time-step to 1
temporarily for these test runs, and 1995 start date changed to 2000.

Hoping this is clear!