Aging details expanded when making databook
Aging is now treated as an equal to the other migrations within the
'transfer details' sheet, even though it is always present even if
num_migrations = 0.
The only other significant change is that a complicated Excel equation
updates the assumption based on if the transmitting population has both
a minimum and maximum age.
So, for example, activating an aging transition from Pop 2 to Pop 4,
where Pop 2 has minimum age 3 and maximum age 7, will produce an
assumption of 0.2 for the transition.
If ages are inaccurately noted, the assumption defaults to 0.