Expanded aging details loaded into model
Surprisingly, once a transfer-details block was set up for aging in the
previous commit, the only thing needed to propagate user values forward
into the model was to take out special-treatment code for aging.
Validation will still be required to make sure aging is a reserved term.
Maybe it should be fixed in Settings.
Regardless, it is now possible to vary population transfers based on
aging, which should fine-tune demographic dynamics, especially for
countries with significant demographic shifts.