Halfway through parameters handling special rules
Parameters that are tagged with special rules are treated as parameters
with custom functions, so that they are handled during
updateDependencies().
If the rule is trivial then there is minimal performance loss. If the
rule is more involved, then they can be updated within the order of
dependencies defined in the cascade worksheet.
Either way, a parameter tagged with special rules maintains a vals_old
backup of values before the application of any special rules, e.g. in
case it is desired to examine population infectiousness prior to
contact-averaging.