Special rules applied only once at start
Dependencies need to be updated twice before a model starts running;
first to potentially calculate junction popsizes and then secondly to
update once those junctions have been propagated.
Special rules should not be iterated more than once before each timestep
(lest weighted-averages be applied to weighted-averages), so this commit
restricts that.
Of course, initialisation logic becomes tricky in the case of junction
outflows with (or dependent on parameters with) special rules, where
special rules should be processed before junctions.
This scenario is unimportant here, given that weighted-averaging is
primarily intended for infectiousness, but this situation can be
reassessed later for rigour and the sake of completeness.