Replacement values updated
If a characteristic is normalised by a zero value, it should not be
restricted from being infinite, even if the flow on effects show up in
plots.
On the other hand, for contacts, a weighted sum of zero implies all
pop_counts involved are zero, which means that the specially-averaged
value for a parameter can be zero.