Node outlinks are now tuples
In preparing for a node to store ids for all links that flow in, it is
worth noting that transfers can originate in other populations.
So inlinks will need to be tuples, where the first element is that of
the ModelPopulation and the second element is that of the link within
the ModelPop.
In that case, might as well be consistent with outlinks, even if they
all originate in the current ModelPop.
All references to Node.outlink_ids thus far have been internal to
model.py or the renamed plotAllOutflows() plotting.py function, so this
change should hopefully have been updated everywhere by this commit.