Custom Titles
Flow plots are done per comp_label and per pop_label.
In the absence of pop_labels, all pops are included in the loop.
In the absence of comp_labels, no compartments are included in the loop.
User can input comp_titles and pop_titles lists of equivalent sizes to
the corresponding labels.
Plot titles are appropriate combinations of comp_title plus pop_title.