Minimalistic program data loaded
Project.data['meta']['progs'] now stores label-name relations.
Project.data['progs'] stores cost and coverage data with associated
years, as well as formats and target pops.
Because years, cost and coverage are all lists, missing data is replaced
by assumptions or np.nan.
This covers most cases of abnormal user entry, but can still give the
unusual situation of no cost-coverage pairs, e.g. if the user puts a
coverage of 10 for 2000 and a cost of 20 for 2005. Validation for this
case has not been decided yet.
Next up is converting this data into a simple progset.