Budget Scenario red herring
After coding up a framework for budget scenario, it struck me that
overwriting the parset at the outset will not do for impacts that are
dynamically generated, e.g. coverages that are weighted across multiple
compartments or populations.
However, I am leaving the Budget Scenario edits rather than reverting,
as they currently function as a Parameter Scenario that ignores scenario
values.
This will make it easier to develop once the entry point for a progset
into model.py is determined.