Passing progset and options into model
Abandoning the scenario mentality for the moment, regarding budgets, the
Project.runSim() function has been extended to take in a progset and
options dictionary.
The idea is that if options['progs_start'] is a year, parameter
overwrites by progset values should start coming into effect.
However, this has so many issues at the current stage, even for
hard-coded parameter targetting, not least of which is the fact that
formats can change from fraction to number.
This preliminary commit is going ahead so as to try and update cascade
and databook with Minsk version before further experimenting.