Format conversion during prog-based overwrite
When impact is derived from a program, to be used to overwrite a
parameter value, format conversion must be applied, if appropriate.
This seems to be working fine, although everything else is still
hard-coded.
Much more work to be done.
Incidentally, the model as is currently maxes out diag, treatment and
success rates. Interestingly, plots show infections still exist, but
this number grows smaller for smaller dt.
This indicates that lag will always result in new diagnoses and newly
treated having to wait a timestep before being treated or cured.