Logic for allocations overwrites sorted out
The options dictionary passed to runSim() now allows for a
'saturate_with_default_budgets' boolean tag, although the name may be
changed later.
If true, any unspecified progset budgets are given default (i.e.
extrapolated) budgets. If the tag is false or non-existent, then the
relevant overwrite process is avoided, such that the parameter follows
calibrated values.
For budget scenarios, this tag should generally be False, because an
analysis is likely to perturb the calibration by funding one program.
For optimisations, the tag may be set to True, although the entire
allocation should be specified in practice.