Handled zero divided by zero characteristics update
Characteristics that have other characteristics as denominators can and
should become infinite if appropriate, e.g. when comparing MDR/DS cases
as a proportion.
However, if the numerator and denominator are both zero, e.g. for
non-comparison 'prevalence' characteristics when populations are empty,
it is more realistic to say that there is a zero prevalence of something
within a zero-size population.
This commit makes sure a numerator of zero takes precedence, hopefully
avoiding any more infs making their way into parameters that use
characteristics in their function legitimately.