Program attributes loaded into data
Given that print conditions are used to blank out lines for program data
entry, a new tag had to be introduced for makeValueEntryArrayBlock(), so
as to allow pre-opened Excel files to be loaded with assumptions and not
blank values.
In any case, it is now possible for the user to provide attribute values
and propagate them into the project data structure.
Pushing them to the progset and then using their values to update impact
parameters in model.py is next.