Initial attribute successes
Attribute functions for impact parameters, used to rescale coverages
when overwriting flow rates, appear to be successfully propagated to the
progset/model and calculated.
Extra features need to be developed, such as coverage disaggregation
over smear as well as modality interactions necessary for Xpert, and
validation/QA is obviously severely limited.
However, this is reasonably the core functionality of programs in place.