Smear-disaggregated coverage handled
The cascade sheet now allows for an 'Impact Groupings' column, allowing
for different impact parameters to be associated with the same tags.
Once in model.py's updateValues() function, the normalisation value for
coverage now adds on source-compartment sizes for any parameters that
are grouped with the impact parameter in focus.
So, for instance, where a coverage of 1000 would be used to describe the
treatment uptake of both Smear Positive and Smear Negative people, that
coverage can now be split proportionally between the two groups by
marking 'spdyes_rate' and 'spmyes_rate' by the same impact grouping tag,
e.g. "a".
Note: This commit also makes a minor fix to progset, storing the label
for a program type within each program, not the name.