Default constraints implemented
It currently takes about 2 minutes to run 7 optimisation iterations for
the Belarus project.
If the optimizer tests the same allocation, the expensive model run
should be averted in objective calculation, but this is yet to be
verified.
Either way, massive performance investigations are required, but we may
need to forego that for the upcoming deadline and rely on
multiprocessing boosts.
In any case, this appears to be the absolute minimalist version of
optimisation. Lots more work required to make it flexibly usable.