Objective extracted for a particular year
Objectives in the options dictionary can be associated with a 'year'
key, e.g. options['objectives']['alive'] = {'weight':-1, 'year':2030.0).
In this case, the objective value is extracted from a year, not summed
across a range (and scaled by dt).
The toc function in utils also currently returns an elapsed time value,
just in case logger.debug() is not printing to screen.