Fixed cost programs seemingly handled
If a program type is marked 'cost_only' in the cascade.xlsx, within a
'Special Tag' column, it only prints a user-entry section for spending
in the data book.
When loading a fixed cost program from the databook, coverages are left
as NaNs and cov_format is None.
Fixed cost programs are typically given no impact parameters (although
this should be verified), so they have no impact in model.py.
They also have appropriate coverage-budget conversion failures and
return an impact of zero.