Rely on results extraction for objective
Objectives keyed in options['objectives'] may have a 'year' key. If this
year exists, values are extracted for the objective only for this year.
Otherwise, values are extracted for the objective in a cumulative manner
from the time that programs are switched on to the end of the
simulation.
This all relies on Results.returnValueAt() to return properly calculated
objectives. As of this time, there is still a dt bug intrinsic to them.