Allow subset optimisations
Locking out programs via constraints is a way to exclude programs from
optimisations, but still account for them in plots.
Alternatively, if 'saturate_with_default_budgets' is set as False, then
only the progset subset actually in 'init_alloc' will be used in the
optimisation.
However, defaultOptimOptions() typically sets that flag as True. If this
is passed to optimizeFunc(), then a full progset init_alloc will be
generated and used for the entirety of the optimisation.