Recursive constraint application
The constrainAllocation() function now appears to handle limit
constraints stored in options['constraints']['limits'][program]['vals'].
When ASD chooses a new allocation, all allocation values beyond the
limits are set to the limits, taking note of which programs hit upper
limits and which hit lower limits.
If the allocation then needs to be scaled up to reach the preferred
total, only programs that haven't hit the upper limit are scaled up (and
by an appropriate amount). A similar situation occurs in the opposite
direction.
This process is repeated recursively until a limit is hit, or until
there is no rescale necessary.