Hotfix for when no previous allocation
Previous allocation in algorithm_refs is used to store the constrained
allocation selected from the previous iteration of ASD for two
reasons...
1. To avoid wasting time on repeating a model run for an identical
allocation.
2. To revert to if constraining is a failure.
However, if the 1st iteration of ASD fails at constraining, there is no
previous allocation to revert to.
Hence the initial previous_alloc in this commit is now storing the
initial allocation.
This should avoid bugs with case 2 while safeguarding case 1.