Minor hotfix
An if statement was in the wrong location, placed between an if-else
fork.
Amusingly, it had no adverse affect on functionality. Checking that an
allocation component was over a limit would still proceed as normal,
because the erroneous if statement would rarely be triggered, relating
to np.nans.
All the same, it's best that this correction takes place for clarity's
sake.