Fix for zero-sum allocations
If an allocation and/or the modifiable parts of it sum up to zero, this
means they are all zero and will never be rescaled through
multiplication.
Therefore, this commit ensures none of the relevant sums are zero by
adding 1 to each budget should the case arise.
This should allow users to optimise init_allocs of zeros.
Obviously the late change in allocation will potentially invalidate
limit constraints, but it does not seem at the moment like there would
be a case where this changed alloc would not be reconstrained in a
subsequent recursive iteration.