Constrain allocation before ASD
ASD does not handle stepsizes well if all budgets passed in are zero.
Hence might as well constrain allocation ahead of time.
Doing this caused a numpy casting problem, presumably because the
allocation of zeros was treated as an array of ints, so this has been
forced to be cast as floats in constrainAllocation().