Initial max-yearly-change constraint introduced
As seen in defaults.py, thre is a new type of constraint keyed by
'max_yearly_change' and then program label, which has a scalar 'val' and
a boolean 'rel'.
This is the first constraint implemented that is actually applicable to
budget scenarios, as it affects updateValues() in model.py.
Effectively, if there is any difference between the intended budget and
the default budget (i.e. the one assumed to correspond with the
calibration), changes are linearly constricted by a derivative, either
relative or absolute.
For instance, {'val':0.5, 'rel':True} should imply that a program with
current funding $1e6 can only gain or shed $5e5 per year in its linear
progression towards intended funding.
Of course, $0 default budgets have issues with relative constraints, so
it is currently up to the scripter to either change val to np.inf or
convert the constraints to something absolute, i.e. 'rel' is false.