Fix to the ramp constraint check
Ramp logic only should be processed if the average derivative (not
multiple) across a number of years since starting programs is greater
than the yearly constraint. (Remember that there is no reason not to
allow the maximal change towards the intended allocation.)
This commit fixes that check so that the zeroth timestep after program
activation does not ignore the constraint.