Hotfix for ramp constraint logic
Previously, it was possible for a relative ramp constraint of 1e12,
which should cause a crash when operating on a default budget of $0 (due
to no change possible), to skip the crash because it would still be
considered greater than the absolute difference between default and
intended budgets.
This commit avoids that by partitioning the derivative check between
absolute and relative constraints, making it specific to each case,
while also moving the zero-budget relative constraint complaint outside
of the derivative check.