Hotfix for summed rates
When rates for multiple populations are extracted within
plotScenarioFlows(), only the rates for the first population were
stored.
This commit sums the np.array objects in the list together, leaving
singular arrays untouched (e.g. that of PLHIV), while combining multiple
arrays together (e.g. that of the general populations).
This is a risky commit though as it is unclear why the first set of
rates was selected to begin with. Perhaps it has something to do with
case usage for another arg being treated as a set.
Thus the alternative has been commented out for easy reversion if
required.