Hotfix to handle duplicate parameter extraction
This commit modifies results.getFlow() so that, if a transition tag is
passed in that maps to multiple links in a ModelPop (e.g. untreated
deaths), the flow rates extracted are summed across the links.
This of course assumes that the stored link values are of the same
length and so on.
Thus untreated death rates now takes into account both undiagnosed and
diagnosed people dying.