Construct the ramped alloc array
When generating Model.prog_vals for the first time, the presence of
ramped constraints generates a cost array for the entire time domain
that starts at a default allocation and then linearly ramps to the
intended initial allocation once programs start.
This should work for an init_alloc that is coverage-based as well.