Parameters as result outputs
If a parameter in a cascade.xlsx file is marked with any non-empty
string in an 'Output' column, it will be treated as a model dependency.
This means it will be calculated as part of the Model.updateValues()
process.
It will then be listed as an output as part of Model.calculateOutputs()
along with all characteristics.