Minor fix to avoid redundant parameters
All parameters marked with a tag are stored as ModelPopulation links.
All others are already marked as dependencies.
So if a transition is marked as an 'Output' in a cascade.xlsx file,
values need to be propagated appropriately from link storage to results,
rather than from dependency storage.
Note: It strikes me that, while this does allow for parameters to be
stored as result outputs, it does not necessarily reflect the true flows
within a model after rescaling. Must consider alternatives.