Improved querying of result values
Mimicked Results.getValueAt() with new and improved method
Results.getValuesAt(), which will hopefully replace the original when
possible, either by renaming previous references or making the former a
special-case wrapper of the latter.
Results.getValuesAt() returns an array of transition flow,
characteristic (including output parameter) or compartment values
between a start and end year, inclusive of the former and exclusive of
the latter.
These are point population sizes or, in the case of flows, annualised
rates summed over all or a specified subset of population groups.
Optional boolean argument 'integrated' can be used to convert a series
of annualised transition flows into a net flow within the specified time
period.
Integration is likely less useful for (non-parameter) characteristics
and compartments, but is handled consistently with flow rates, unlike
the original getValueAt() that simply summed point popsizes.