Single plots working again
Following a temporary linestyle fix in plotting.py and
ResultSet.getValuesAt() adaptations, it is possible to plot
characteristics once again.
A lot of the plotting methodology will have to be reviewed and fixed
later, given the _plotLine() calls.
Figure labels will need to be updated, as will parset references.
However, the basic calibration concept works.