Coverage overflow renormalisation
Model.updateValues() now stores a list of individual program impacts
during an impact parameter loop before summing them at the end.
It also calculates and lists 'overflow', namely how much greater a
program's coverage is versus the number of people that can be covered
(a.k.a. source set size).
At the end of the impact parameter iteration, overflow is summed across
relevant programs.
If it exceeds 1, then all impacts are renormalised under the assumption
that a program with overflow 5 has 10 times better outreach than a
program with overflow 0.5 (i.e. the one with a greater coverage is more
accessible).
So far this appears to limit impact appropriately, though more thorough
testing may be required.